RDKBDEV-2100 : Xconf script bug fix to avoid multiple script instances
Reason for change: When Periodic FW check is toggled multiple times, It was observed that previous process
was in sleep and new process got started for Firmware download, creating multiple script execution.
To avoid this, check before starting main app if there is already any xconf download process under execution,
if so, kill it and proceed.

Test Procedure: Test results are captured in the ticket

Risks: None

Change-Id: Id86045f9ba68ebcfca94d5e752921d109435b9d5
Signed-off-by: Mohamed Shaikh <mohamed.shaikh@t-systems.com>